+++ src/bin/nkscan/main.rs
@@ -6,6 +6,7 @@ use tracing_subscriber::EnvFilter;
 
 mod cancel;
 mod cli;
+mod connect;
 mod dump;
 mod eject;
 mod io;